+++ app/lib/ui/components/components.dart
@@ -69,8 +69,14 @@ List<Component> components(BuildContext context) {
       AdaptiveImageHelper.getImage(context, AppAssets.images.ilComponentsChip),
       context.l10n.app_components_chip_description_text,
       [
-        VariantComponent(context.l10n.app_components_filterChip_label, ChipFilterDemoScreen()),
-        VariantComponent(context.l10n.app_components_suggestionChip_label, ChipSuggestionDemoScreen()),
+        VariantComponent(
+            context.l10n.app_components_filterChip_label,
+            ChipFilterDemoScreen()
+        ),
+        VariantComponent(
+            context.l10n.app_components_suggestionChip_label,
+            ChipSuggestionDemoScreen()
+        ),
       ],
     ),
     Component.withVariant(