+++ ouds_core/lib/components/chip/ouds_filter_chip.dart
@@ -37,8 +37,7 @@ enum OudsChipStyle {
   selected,
 }
 
-//// [OUDS Filter Chip design guidelines](https://unified-design-system.orange.com/472794e18/p/73c701-components)
-/// OUDS Filter Chip design guidelines will be available soon
+// TODO: Add documentation URL once it is available
 ///
 /// A filter chip is a compact UI element used in a design system to represent a filter option that can be selected or deselected by the user.
 /// Filter chips allow users to refine content or data by applying one or more filters in a visually accessible and interactive way.